feat(ModalForm): fix test case
